+++ CHANGES.md
@@ -17,6 +17,8 @@ Bug Fixes
 * Fix crash in direct mode callbacks with certain type conversions - [@twall](https://github.com/twall).
 * More thoroughly propagate unexpected exceptions generated in jnidispatch - [@twall](https://github.com/twall).
 * Cleanup maven poms and publishing to central repo - [@bhamail](https://github.com/bhamail).
+* [#129](https://github.com/twall/jna/issues/129): Allow `Memory` field in structure - [@twall](https://github.com/twall).
+* Preserve `PointerType` fields on `Structure.read()` if unchanged - [@twall](https://github.com/twall).
 
 Release 3.4.2
 =============